python-backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
flask-cors
google-generativeai
# Add other Python packages you need here, for example:
  # For PostgreSQL connection
requests         # For making HTTP requests to external APIs

    
